tests were simplified
