+++ src/providers/data/api_dataprovider.py
@@ -1,12 +1,12 @@
-# Data for testing LoginPage Page
+# Data for testing API Calls
 from src.config.config import CFG
 from src.helpers.enums import ApiResponseType
 
 api_data = [
-    ({"description": "Login with invalid credentials",
+    ({"description": "API call Dashboard with invalid credentials",
       "email": "valid@ema.il", "password": "secret_password",
       "expected_result": ApiResponseType.error}),
-    ({"description": "Login with valid credentials",
+    ({"description": "API call Dashboard with valid credentials",
       "email": CFG.USER_EMAIL, "password": CFG.USER_PASSWORD,
       "expected_result": ApiResponseType.ok}),
 ]
