--- src/tests/conftest.py
@@ -1,4 +1,0 @@
-from src.tests.fixtures.fixture_config import *
-from src.tests.fixtures.fixture_logging import *
-from src.tests.fixtures.fixture_login_page import *
-from src.tests.fixtures.fixture_api import *
