+++ tests/test_web_login_page.py
@@ -10,15 +10,11 @@ class TestWebLoginPage:
 
     @mark.order(1)
     def test_login(self, web_driver, logger, dp_login_page):
-        logger.info(f'Test case for login with email {dp_login_page["email"]}')
+        logger.info(f'Test case for email {dp_login_page["email"]}')
         login_url = get_url(CFG.BASE_URL, CidEndpoints.login)
         web_driver.get(login_url)
         login_page = LoginPage(web_driver)
         login_page.login(dp_login_page["email"], dp_login_page["password"])
         current_url = web_driver.current_url
-        logger.info(dp_login_page["email"])
-        logger.info(dp_login_page["password"])
-        logger.info(login_url)
-        logger.info(current_url)
         assert ((current_url != login_url) == dp_login_page["expected_result"])
         logger.info("Test PASSED")
